Move db connection factory interface to application layer
