+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/ActorProvider.cs
@@ -16,7 +16,7 @@ using System;
 using System.Threading.Tasks;
 using Dapper;
 using Energinet.DataHub.Core.App.Common.Abstractions.Actor;
-using Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess;
+using Energinet.DataHub.MeteringPoints.Application.Common;
 using Microsoft.Extensions.Logging;
 
 namespace Energinet.DataHub.MeteringPoints.Infrastructure
