+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/DataAccess/EnergySuppliers/Queries/EnergySuppliersByMeteringPointIdQueryHandler.cs
@@ -17,6 +17,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Dapper;
+using Energinet.DataHub.MeteringPoints.Application.Common;
 using Energinet.DataHub.MeteringPoints.Application.EnergySuppliers;
 using Energinet.DataHub.MeteringPoints.Application.EnergySuppliers.Queries;
 using MediatR;
