+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/DataAccess/MeteringPoints/Queries/MeteringPointByIdQueryHandler.cs
@@ -16,6 +16,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Dapper;
+using Energinet.DataHub.MeteringPoints.Application.Common;
 using Energinet.DataHub.MeteringPoints.Application.Queries;
 using MediatR;
