+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/InternalCommands/InternalCommandProcessor.cs
@@ -16,8 +16,8 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Dapper;
+using Energinet.DataHub.MeteringPoints.Application.Common;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
-using Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess;
 using Energinet.DataHub.MeteringPoints.Infrastructure.Serialization;
 using Microsoft.Extensions.Logging;
 using Polly;
