+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/Providers/MeteringPointOwnership/MeteringPointOwnershipProvider.cs
@@ -15,9 +15,9 @@
 using System;
 using System.Threading.Tasks;
 using Dapper;
+using Energinet.DataHub.MeteringPoints.Application.Common;
 using Energinet.DataHub.MeteringPoints.Application.Providers.MeteringPointOwnership;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
-using Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess;
 
 namespace Energinet.DataHub.MeteringPoints.Infrastructure.Providers.MeteringPointOwnership
 {
