+++ source/Energinet.DataHub.MeteringPoints.IntegrationTests/AssertPersistedMeteringPoint.cs
@@ -14,12 +14,12 @@
 
 using System;
 using Dapper;
+using Energinet.DataHub.MeteringPoints.Application.Common;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components.Addresses;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components.MeteringDetails;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
-using Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess;
 using Xunit;
 
 namespace Energinet.DataHub.MeteringPoints.IntegrationTests
